stupid crypt
Just tried to implement a simple Caesar crypto, it's really basic, but it -at least- have a keygen function that secures a bit of securing of files. It doesn't delete plain files, so it's pretty useless and I haven't implemented any algorithm to loop every file in a directory, nor a way to select a set of files.